% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.example_out_table.R
\docType{data}
\name{example_out_table}
\alias{example_out_table}
\title{example_out_table:}
\format{
A data frame
}
\usage{
data(example_out_table)
}
\description{
otu_table example for taxon-specific lysis estimation
}
\examples{
data(example_out_table)
}
\keyword{example_out_table}
