% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lysis.R
\name{lysis}
\alias{lysis}
\title{A lysis Function}
\usage{
lysis(otu_table)
}
\arguments{
\item{otu_table}{'Path/To/Your/OTU_TABLE_FILE'; For example: otu_table="/home/kevin/Desktop/data/10reads-plus/otu_table_qiime2_otu_16S.txt".}
}
\description{
This function allows calculating the taxon-specific lysis of microbe, based on the input of a tabular-separated otu table (.txt) that contains data from three rDNA/rRNA fractions (Cellular rDNA, Cellular rRNA and Extracellular rRNA) of each water sample.
}
\examples{
##### If you want to estimate the taxon-specific lysis for prokaryotes, then run
lysis(otu_table="Path/To/Your/OTU_TABLE_FILE_16S.txt")











}
\keyword{lysis}
