% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.metadata.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{metadata:}
\format{
A data frame
}
\usage{
data(metadata)
}
\description{
metadata example for taxon-specific lysis rate estimation
}
\examples{
data(metadata)
}
\keyword{metadata}
