% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.quadra_16S.R
\docType{data}
\name{quadra_16S}
\alias{quadra_16S}
\title{quadra_16S:}
\format{
A data frame
}
\usage{
data(quadra_16S)
}
\description{
otu_table example for taxon-specific lysis estimation
}
\examples{
data(quadra_16S)
}
\keyword{quadra_16S}
